(feature) Display username and profile picture
- Create SocialAuthUsersocialauth model to query table
social_auth_usersocialauth (which is created by python-social-auth)
so as to provide a means of retrieving the facebook uid
- Refactor app.views to display additional username and uid information
for logged in users
- Update url mapping of app.views according to PR nitpick recommendation
- Update AuthService's url reference to loginStatus url
- Fetch URL of the currently logged in user if the call to AuthService
resolves successfully. This URl is being stored in 'profile_picture'
scope variable. The username from this call is also being stored in
'username' scope variable
- The dashboard has two elements that refer to the 'profile_picture'
and 'username' scope variable. These values are displayed on the dash-
board view
- Create a materialize preloader for uploading photos on the dashboard