(fix) Create foreign key in Preview model
- Create a foreign key reference to Photo model in the Preview model.
This was done to ensure that each preview has an associated uploaded
photo
- Refactor the serializer to require the photo as a read/write attribute
and change the path attribute to a read only field. This was done to
eliminate redundancy (since with the photo object, a path can be obtained)
- Refactor the create method of the PreviewViewSet to fetch a Photo
object from the database and use this object as a foreign key to all
Preview objects. The Photo object is also used to provide Preview's path
attribute with a value.
- Refactor the DashboardController to send the photo_id as the modified
backend requires
- Fix failing tests as a result of changes to the back end